fix-delete-user

+++ be/src/controllers/admin.controller.ts
@@ -8,6 +8,7 @@ import { IAdmin } from '../models/admin';
 import postRepository from '../repositories/post.repository';
 import productRepository from '../repositories/product.repository';
 import attribute_productRepository from '../repositories/attribute_product.repository';
+import OrderRepo from '../repositories/order.repository';
 
 interface MulterRequest extends Request {
     files: Express.Multer.File[];
@@ -143,7 +144,10 @@ class AdminController {
     async deleteUser(req: Request, res: Response): Promise<void> {
         try {
             const id = req.params.id;
-
+            if (await OrderRepo.userIsOrdering(id)) {
+                res.status(400).send('Bạn không thể xóa user này, user đang trong quá trình nhận đơn hàng');
+                return;
+            }
             await Promise.all([
                 UserRepo.deleteUser(id),
                 postRepository.deletePostsByUserId(id),
